+++ src/main/java/roomescape/entity/Reservation.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.entity;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
@@ -8,7 +8,6 @@ public class Reservation {
     private final String name;
     private final LocalDate date;
     private final LocalTime time;
-
 
     public Reservation(Long id, String name, LocalDate date, LocalTime time) {
         this.id = id;
